move .with_attributes logic to shared lib
